Added utility function to remove a resource from a FHIR bundle
