+++ prime-router/src/main/kotlin/fhirengine/utils/FHIRBundleHelpers.kt
@@ -2,12 +2,19 @@ package gov.cdc.prime.router.fhirengine.utils
 
 import gov.cdc.prime.router.CustomerStatus
 import gov.cdc.prime.router.Receiver
+import gov.cdc.prime.router.fhirengine.translation.hl7.utils.FhirPathUtils
 import io.github.linuxforhealth.hl7.data.Hl7RelatedGeneralUtils
+import org.hl7.fhir.r4.model.Base
 import org.hl7.fhir.r4.model.Bundle
+import org.hl7.fhir.r4.model.DiagnosticReport
 import org.hl7.fhir.r4.model.Endpoint
 import org.hl7.fhir.r4.model.Identifier
+import org.hl7.fhir.r4.model.Observation
+import org.hl7.fhir.r4.model.Property
 import org.hl7.fhir.r4.model.Provenance
 import org.hl7.fhir.r4.model.Reference
+import java.util.stream.Stream
+import kotlin.streams.toList
 
 /**
  * A collection of helper functions that modify an existing FHIR bundle.
@@ -50,5 +57,109 @@ object FHIRBundleHelpers {
         }
 
         if (targetList.isNotEmpty()) provenanceResource.target.addAll(targetList)
+    }
+
+    /**
+     * Gets all properties for a [Base] resource recursively and filters only its references
+     *
+     * @return a list of reference identifiers for a [Base] resource
+     *
+     */
+    fun Base.getResourceReferences(): List<String> {
+        return filterReferenceProperties(this.getResourceProperties())
+    }
+
+    /**
+     * Gets all properties for a [Base] resource recursively
+     *
+     * @return a list of all [Property] for a [Base] resource
+     */
+    fun Base.getResourceProperties(): List<Property> {
+        return this.children().stream().flatMap { getChildProperties(it) }.toList()
+    }
+
+    /**
+     * Filters the [properties] by only properties that have a value and are of type [Reference]
+     *
+     * @return a list containing only the references in [properties]
+     */
+    private fun filterReferenceProperties(properties: List<Property>): List<String> {
+        return properties.filter { it.hasValues() }.flatMap { it.values }
+            .filterIsInstance<Reference>().map { it.reference }
+    }
+
+    /**
+     * Gets all child properties for a resource [property] recursively
+     *
+     * @return a flatmap stream of all child properties on a [property]
+     */
+    private fun getChildProperties(property: Property): Stream<Property> {
+        return Stream.concat(
+            Stream.of(property),
+            property.values.flatMap { it.children() }.stream().flatMap { getChildProperties(it) }
+        )
+    }
+
+    /**
+     * Gets all diagnostic report that have no observations from a [bundle]
+     *
+     * @return a list of [Base] diagnostic reports that have no observations
+     */
+    fun getDiagnosticReportNoObservations(bundle: Bundle): List<Base> {
+        return FhirPathUtils.evaluate(
+            null,
+            bundle,
+            bundle,
+            "Bundle.entry.resource.ofType(DiagnosticReport).where(result.empty())"
+        )
+    }
+
+    /**
+     * Deletes a [resource] from a bundle, removes all references to the [resource] and any orphaned children.
+     * If the [resource] is a [DiagnosticReport] and it has no observations. The [DiagnosticReport] and all
+     * its references will be removed
+     *
+     */
+    fun Bundle.deleteResource(resource: Base) {
+
+        if (this.entry.find { it.fullUrl == resource.idBase } == null) {
+            throw IllegalStateException("Cannot delete resource. FHIR bundle does not contain this resource")
+        }
+
+        // Get the resource children references
+        val resourceChildren = resource.getResourceReferences()
+        // get all resources except the resource being removed
+        val allResources = this.entry.filter { it.resource.id != resource.idBase }
+        // get all references for every resource
+        val allReferences = allResources.flatMap { it.getResourceReferences() }
+
+        // remove orphaned children
+        resourceChildren.forEach { child ->
+            if (!allReferences.contains(child)) {
+                this.entry.removeIf { it.resource.id == child }
+            }
+        }
+
+        // Go through every resource and check if the resource has a reference to the resource being deleted
+        // if there is remove the reference
+        this.entry.forEach { res ->
+            res.resource.getResourceProperties().forEach { property ->
+                property.values.forEach {
+                    if (it is Reference && it.reference == resource.idBase) {
+                        it.reference = null
+                        it.resource = null
+                    }
+                }
+            }
+        }
+
+        // if resource being removed is an observation
+        // check if there are any diagnostic reports with no observations and remove them
+        if (resource is Observation) {
+            getDiagnosticReportNoObservations(this).forEach { this.deleteResource(it) }
+        }
+
+        // finally remove the resource from the bundle
+        this.entry.removeIf { it.fullUrl == resource.idBase }
     }
 }
