+++ prime-router/src/test/kotlin/fhirengine/engine/fhirRouterTests/FHIRBundleHelpersTests.kt
@@ -1,0 +1,344 @@
+package gov.cdc.prime.router.fhirengine.engine.fhirRouterTests
+
+import assertk.assertThat
+import assertk.assertions.isEqualTo
+import assertk.assertions.isFalse
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import assertk.assertions.isNull
+import assertk.assertions.isTrue
+import gov.cdc.prime.router.ActionLogger
+import gov.cdc.prime.router.CustomerStatus
+import gov.cdc.prime.router.DeepOrganization
+import gov.cdc.prime.router.Metadata
+import gov.cdc.prime.router.Organization
+import gov.cdc.prime.router.Receiver
+import gov.cdc.prime.router.Schema
+import gov.cdc.prime.router.Topic
+import gov.cdc.prime.router.azure.BlobAccess
+import gov.cdc.prime.router.azure.DatabaseAccess
+import gov.cdc.prime.router.azure.QueueAccess
+import gov.cdc.prime.router.fhirengine.engine.RawSubmission
+import gov.cdc.prime.router.fhirengine.translation.hl7.utils.CustomContext
+import gov.cdc.prime.router.fhirengine.translation.hl7.utils.FhirPathUtils
+import gov.cdc.prime.router.fhirengine.utils.FHIRBundleHelpers
+import gov.cdc.prime.router.fhirengine.utils.FHIRBundleHelpers.deleteResource
+import gov.cdc.prime.router.fhirengine.utils.FHIRBundleHelpers.getResourceProperties
+import gov.cdc.prime.router.fhirengine.utils.FHIRBundleHelpers.getResourceReferences
+import gov.cdc.prime.router.fhirengine.utils.FhirTranscoder
+import io.mockk.clearAllMocks
+import io.mockk.mockkClass
+import io.mockk.spyk
+import org.hl7.fhir.r4.model.Endpoint
+import org.hl7.fhir.r4.model.Observation
+import org.hl7.fhir.r4.model.Provenance
+import org.jooq.tools.jdbc.MockConnection
+import org.jooq.tools.jdbc.MockDataProvider
+import org.jooq.tools.jdbc.MockResult
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.TestInstance
+import java.io.File
+import java.lang.IllegalStateException
+import java.util.UUID
+import kotlin.test.Test
+import kotlin.test.assertFailsWith
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class FHIRBundleHelpersTests {
+    val dataProvider = MockDataProvider { emptyArray<MockResult>() }
+    val connection = MockConnection(dataProvider)
+    val accessSpy = spyk(DatabaseAccess(connection))
+    val blobMock = mockkClass(BlobAccess::class)
+    val queueMock = mockkClass(QueueAccess::class)
+    val metadata = Metadata(schema = Schema(name = "None", topic = Topic.FULL_ELR, elements = emptyList()))
+    val bodyUrl = "http://anyblob.com"
+    private val defaultReceivers = listOf(
+        Receiver(
+            "full-elr-hl7",
+            "co-phd",
+            Topic.FULL_ELR,
+            CustomerStatus.ACTIVE,
+            "one"
+        ),
+        Receiver(
+            "full-elr-hl7-2",
+            "co-phd",
+            Topic.FULL_ELR,
+            CustomerStatus.INACTIVE,
+            "one"
+        )
+    )
+    val oneOrganization = DeepOrganization(
+        "co-phd",
+        "test",
+        Organization.Jurisdiction.FEDERAL,
+        receivers = defaultReceivers
+    )
+
+    val message = spyk(RawSubmission(UUID.randomUUID(), "http://blob.url", "test", "test-sender"))
+
+    private val validFhirWithProvenance = """
+    {
+        "resourceType": "Bundle",
+        "id": "1666038428133786000.94addcb6-835c-4883-a095-0c50cf113744",
+        "meta": {
+        "lastUpdated": "2022-10-17T20:27:08.149+00:00",
+        "security": [
+            {
+                "code": "SECURITY",
+                "display": "SECURITY"
+            }
+        ]
+    },
+        "identifier": {
+        "value": "MT_COCAA_ORU_AAPHELR.1.6214638"
+    },
+        "type": "message",
+        "timestamp": "2028-08-08T15:28:05.000+00:00",
+        "entry": [
+            {
+                "fullUrl": "Provenance/1666038430962443000.9671377b-8f2b-4f5c-951c-b43ca8fd1a25",
+                "resource": {
+                    "resourceType": "Provenance",
+                    "id": "1666038430962443000.9671377b-8f2b-4f5c-951c-b43ca8fd1a25",
+                    "recorded": "2028-08-08T09:28:05-06:00",
+                    "activity": {
+                        "coding": [
+                            {
+                                "system": "http://terminology.hl7.org/CodeSystem/v2-0003",
+                                "code": "R01",
+                                "display": "ORU_R01"
+                            }
+                        ]
+                    }
+                }
+            }
+        ]
+    }
+    """
+
+    @BeforeEach
+    fun reset() {
+        clearAllMocks()
+    }
+
+    @Test
+    fun `test adding receivers to bundle`() {
+        // set up
+        val bundle = FhirTranscoder.decode(validFhirWithProvenance)
+        val receiversIn = listOf(oneOrganization.receivers[0])
+
+        // act
+        FHIRBundleHelpers.addReceivers(bundle, receiversIn)
+
+        // assert
+        val provenance = bundle.entry.first { it.resource.resourceType.name == "Provenance" }.resource as Provenance
+        val outs = provenance.target
+        val receiversOut = outs.map { (it.resource as Endpoint).identifier[0].value }
+        assert(receiversOut.isNotEmpty())
+        assert(receiversOut[0] == "co-phd.full-elr-hl7")
+    }
+
+    @Test
+    fun `test skipping inactive receivers (only inactive)`() {
+        // set up
+        val bundle = FhirTranscoder.decode(validFhirWithProvenance)
+        val receiversIn = listOf(oneOrganization.receivers[1])
+
+        // act
+        FHIRBundleHelpers.addReceivers(bundle, receiversIn)
+
+        // assert
+        val provenance = bundle.entry.first { it.resource.resourceType.name == "Provenance" }.resource as Provenance
+        val outs = provenance.target
+        val receiversOut = outs.map { (it.resource as Endpoint).identifier[0].value }
+        assert(receiversOut.isEmpty())
+    }
+
+    @Test
+    fun `test skipping inactive receivers (mixed)`() {
+        // set up
+        val bundle = FhirTranscoder.decode(validFhirWithProvenance)
+        val receiversIn = oneOrganization.receivers
+
+        // act
+        FHIRBundleHelpers.addReceivers(bundle, receiversIn)
+
+        // assert
+        val provenance = bundle.entry.first { it.resource.resourceType.name == "Provenance" }.resource as Provenance
+        val outs = provenance.target
+        val receiversOut = outs.map { (it.resource as Endpoint).identifier[0].value }
+        assert(receiversOut.isNotEmpty())
+        assert(receiversOut[0] == "co-phd.full-elr-hl7")
+    }
+
+    @Test
+    fun `Test retrieving references for a Bundle resource`() {
+        val actionLogger = ActionLogger()
+        val fhirBundle = File("src/test/resources/fhirengine/engine/valid_data.fhir").readText()
+        val messages = FhirTranscoder.getBundles(fhirBundle, actionLogger)
+        val bundle = messages[0]
+
+        var diagnosticReport = FhirPathUtils.evaluate(
+            null,
+            bundle,
+            bundle,
+            "Bundle.entry.resource.ofType(DiagnosticReport)[0]"
+        )[0]
+
+        val diagnosticReportReferences = diagnosticReport.getResourceReferences()
+
+        assertThat(diagnosticReportReferences.count()).isEqualTo(5)
+    }
+
+    @Test
+    fun `Test retrieving properties for a Bundle resource`() {
+        val actionLogger = ActionLogger()
+        val fhirBundle = File("src/test/resources/fhirengine/engine/valid_data.fhir").readText()
+        val messages = FhirTranscoder.getBundles(fhirBundle, actionLogger)
+        val bundle = messages[0]
+
+        val diagnosticReport = FhirPathUtils.evaluate(
+            null,
+            bundle,
+            bundle,
+            "Bundle.entry.resource.ofType(DiagnosticReport)[0]"
+        )[0]
+
+        assertThat(diagnosticReport.getResourceProperties()).isNotEmpty()
+    }
+
+    @Test
+    fun `Test find Diagnostic report no observation`() {
+        val actionLogger = ActionLogger()
+        val fhirBundle =
+            File("src/test/resources/fhirengine/engine/bundle_diagnostic_report_no_observations.fhir").readText()
+        val messages = FhirTranscoder.getBundles(fhirBundle, actionLogger)
+        val bundle = messages[0]
+
+        assertThat(FHIRBundleHelpers.getDiagnosticReportNoObservations(bundle).count()).isEqualTo(1)
+    }
+
+    @Test
+    fun `Test Removing observation from diagnostic report with single observation`() {
+        val actionLogger = ActionLogger()
+        val fhirBundle = File("src/test/resources/fhirengine/engine/valid_data.fhir").readText()
+        val messages = FhirTranscoder.getBundles(fhirBundle, actionLogger)
+        val bundle = messages[0]
+
+        val diagnosticReport = FhirPathUtils.evaluate(
+            null,
+            bundle,
+            bundle,
+            "Bundle.entry.resource.ofType(DiagnosticReport)[0]"
+        )[0]
+
+        var observation = FhirPathUtils.evaluate(
+            CustomContext(bundle, bundle),
+            diagnosticReport,
+            bundle,
+            "%resource.result[0].resolve()"
+        )[0]
+
+        val observationReferences = observation.getResourceReferences()
+
+        assertThat(observationReferences.count()).isEqualTo(4)
+
+        // Verify that the bundle contains the Diagnostic Report and observation being removed
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == diagnosticReport.idBase
+            }
+        ).isNotNull()
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == observation.idBase
+            }
+        ).isNotNull()
+
+        bundle.deleteResource(observation)
+
+        // Verify that the bundle no longer contains the Diagnostic Report that contained the observation
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == diagnosticReport.idBase
+            }
+        ).isNull()
+
+        // Verify that the bundle no longer contains the observation deleted
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == observation.idBase
+            }
+        ).isNull()
+
+        // Verify that the resources referenced in this observation got removed
+        // The observation being deleted has four references.
+        // Two are referenced in other resources(Patient and Encounter) and should not be removed.
+        // The other two (Practitioner and Organization) should be removed
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == observationReferences.first { it.contains("Encounter") }
+            }
+        ).isNotNull()
+
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == observationReferences.first { it.contains("Patient") }
+            }
+        ).isNotNull()
+
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == observationReferences.first { it.contains("Organization") }
+            }
+        ).isNull()
+
+        assertThat(
+            bundle.entry.find {
+                it.fullUrl == observationReferences.first { it.contains("Practitioner") }
+            }
+        ).isNull()
+    }
+
+    @Test
+    fun `Test Removing observation from diagnostic report with multiple observations`() {
+        val actionLogger = ActionLogger()
+        val fhirBundle = File("src/test/resources/fhirengine/engine/bundle_multiple_observations.fhir").readText()
+        val messages = FhirTranscoder.getBundles(fhirBundle, actionLogger)
+        val bundle = messages[0]
+        val diagnosticReport = FhirPathUtils.evaluate(
+            CustomContext(bundle, bundle),
+            bundle,
+            bundle,
+            "Bundle.entry.resource.ofType(DiagnosticReport)[0]"
+        )[0]
+
+        val observation = FhirPathUtils.evaluate(
+            CustomContext(bundle, bundle),
+            diagnosticReport,
+            bundle,
+            "%resource.result[0].resolve()"
+        )[0]
+
+        // Check that diagnostic report has multiple observations
+        // and contains the observation being removed
+        assertThat(diagnosticReport.getResourceReferences().count()> 1).isTrue()
+        assertThat(diagnosticReport.getResourceReferences().contains(observation.idBase)).isTrue()
+        messages[0].deleteResource(observation)
+        // Check that the diagnostic report still has multiple observations and
+        // no longer contains the observation removed
+        assertThat(diagnosticReport.getResourceReferences().count()> 1).isTrue()
+        assertThat(diagnosticReport.getResourceReferences().contains(observation.idBase)).isFalse()
+    }
+    @Test
+    fun `Test Removing an observation that doesn't exist`() {
+        val actionLogger = ActionLogger()
+        val fhirBundle = File("src/test/resources/fhirengine/engine/bundle_multiple_observations.fhir").readText()
+        val messages = FhirTranscoder.getBundles(fhirBundle, actionLogger)
+        val observation = Observation()
+        observation.id = "test"
+
+        assertFailsWith<IllegalStateException> { messages[0].deleteResource(observation) }
+    }
+}
